Don't change the output type for np.any/np.all.
This was forcing us to convert the entire source array to boolean and then operating on it.  Huge performance hit.  It was actually preventing the real data type handlers from being executed as everything was processed as boolean.  We convert the results to boolean at the end of the process.